third_party/WebKit/LayoutTests/fast/dom/custom/element-names.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
function createRegisterParameters() {
    return {
        prototype: Object.create(HTMLElement.prototype, { thisIsPrototype: { value: true } })
    };
}

test(function () {
    var invalidNames = [
        // Hyphen missing:
        'foo',
        'xfoo',
        // Reserved names:
        // hyphen-containing names from the applicable specifications,
        // namely the SVG and the MathML.
        'annotation-xml',
        // names containing colon
        ':-xfoo',
        'xfoo-:yfoo'
    ];

    for (var i = 0; i < invalidNames.length; i++) {
        var invalidName = invalidNames[i];
        assert_throws('SYNTAX_ERR', function (){
            document.registerElement(invalidName, createRegisterParameters());
        }, 'register invalid custom element name: ' + invalidName);
    }

}, 'invalid and reserved names');

test(function (){
    document.registerElement('x-foo', createRegisterParameters());
    assert_throws('NOT_SUPPORTED_ERR', function (){
        document.registerElement('x-foo', createRegisterParameters());
    }, 'register already registered "x-foo"');
    assert_throws('NOT_SUPPORTED_ERR', function (){
        document.registerElement('X-FOO', createRegisterParameters());
    }, 'register "X-FOO", same as already registered "x-foo"');
}, 'name conflicts');

test(function () {
    var validNameTagMap = {
        'y-bar': 'Y-BAR',
        'yz-bar': 'YZ-BAR',
        'y-z-bar': 'Y-Z-BAR',
        'y--bar' : 'Y--BAR'
    };
    for(var elementName in validNameTagMap) {
        var expectedTagName = validNameTagMap[elementName];
        var tagName = (new (document.registerElement(elementName, createRegisterParameters()))()).tagName;
        assert_equals(tagName, expectedTagName, tagName + ' must be treated as a custom element');
    }
}, 'strange but valid names');

</script>
